+++ src/NCalc.Core/Factories/ExpressionFactory.cs
@@ -11,13 +11,23 @@ public sealed class ExpressionFactory(
     IEvaluationVisitorFactory? evaluationVisitorFactory = null
 ) : IExpressionFactory
 {
-    public Expression Create(string expression, ExpressionConfiguration? configuration = null)
+    public Expression Create(
+        string expression,
+        ExpressionConfiguration? configuration = null,
+        ExpressionContext? context = null,
+        CultureInfo? cultureInfo = null)
     {
-        return new Expression(expression, configuration, logicalExpressionFactory, cache, evaluationVisitorFactory);
+        return new Expression(expression, configuration ?? new(), context ?? new(),
+            cultureInfo ?? CultureInfo.CurrentCulture, logicalExpressionFactory, cache, evaluationVisitorFactory);
     }
 
-    public Expression Create(LogicalExpression logicalExpression, ExpressionConfiguration? configuration = null)
+    public Expression Create(
+        LogicalExpression logicalExpression,
+        ExpressionConfiguration? configuration = null,
+        ExpressionContext? context = null,
+        CultureInfo? cultureInfo = null)
     {
-        return new Expression(logicalExpression, configuration, logicalExpressionFactory, cache, evaluationVisitorFactory);
+        return new Expression(logicalExpression, configuration ?? new(), context ?? new(),
+            cultureInfo ?? CultureInfo.CurrentCulture, logicalExpressionFactory, cache, evaluationVisitorFactory);
     }
 }